requirements.txt
aiohttp==3.9.1
beautifulsoup4==4.12.2
discord.py==2.3.2
openai>=1.0.0
schedule==1.2.0
